feat: pitch-tracked pulse width via audio-rate PWM oscillator
The real TB-303 varies its square-wave duty cycle with pitch (the pitch CV
drives the PWM input): a wide pulse (~71%) low down, narrowing (~45%) up high.
Open303's oscillator can't reproduce this because its pulse comes from a
fixed-symmetry wavetable whose duty cycle is baked in at render time; changing
it means re-running an FFT/iFFT over the mip-map, far too costly per note, let
alone per sample.

This reworks the oscillator to synthesise the pulse at audio rate:

- Rename rosic_BlendOscillator -> dfl_BlendOscillator (namespace dfl). The phase
  accumulator, mip-map table selection and interpolation remain Robin Schmidt's
  rosic code; the pulse-generation core is new.
- Pulse is now synthesised as -tanh(drive*(saw2 - saw1) + offset) from two
  phase-offset reads of the saw table, so pulse width is a continuous per-sample
  parameter. setPulseWidth() just stores a value now (no wavetable regen).
  Uses rosic::tanhApprox to stay off a real transcendental, consistent with the
  ladder filter's polynomial shaper.
- Drop the now-unused second (square) wavetable; reroute the tanh-shaper
  drive/offset/phase controls from that table to the oscillator.
- Open303::processSample sets pulse width per sample from a log-frequency curve
  (71% at C1 -> 45% at C4). The curve is intentionally not clamped to the
  keyboard range - it continues past the endpoints as the analog CV relationship
  would; only the final pulse width is clamped [25,75] so extreme octaves stay
  musical.

Cost is negligible (monophonic; ~one extra saw read + tanhApprox per sample) and
it removes the per-note FFT spike that a wavetable-regen approach would incur.
