+++ src/dsp/open303/dfl_BlendOscillator.cpp
@@ -1,0 +1,80 @@
+#include "dfl_BlendOscillator.h"
+using namespace dfl;
+using rosic::MipMappedWaveTable;
+
+//-------------------------------------------------------------------------------------------------
+// construction/destruction:
+
+BlendOscillator::BlendOscillator()
+{
+  // init member variables:
+  tableLengthDbl       = (double) MipMappedWaveTable::tableLength;  // typecasted version
+  sampleRate           = 44100.0;
+  freq                 = 440.0;
+  increment            = (tableLengthDbl*freq)/sampleRate;
+  phaseIndex           = 0.0;
+  startIndex           = 0.0;
+  blend                = 0.0;
+  waveTable1           = NULL;
+
+  // audio-rate PWM parameters (two-saw subtraction method):
+  pulseWidth           = 0.5;           // 50% duty cycle default
+  tanhDrive            = dB2amp(36.9);  // 303-style tanh drive
+  tanhOffset           = 4.37;          // 303-style DC offset
+  squarePhaseShift     = 0.0;           // pulse aligned with saw
+
+  // somewhat redundant:
+  setSampleRate(44100.0);          // sampleRate = 44100 Hz by default
+  setFrequency (440.0);            // frequency = 440 Hz by default
+  setStartPhase(0.0);              // startPhase = 0 by default
+
+  setWaveForm1(MipMappedWaveTable::SAW);
+
+  resetPhase();
+}
+
+BlendOscillator::~BlendOscillator()
+{
+
+}
+
+//-------------------------------------------------------------------------------------------------
+// parameter settings:
+
+void BlendOscillator::setSampleRate(double newSampleRate)
+{
+  if( newSampleRate > 0.0 )
+    sampleRate = newSampleRate;
+  sampleRateRec = 1.0 / sampleRate;
+  increment = tableLengthDbl*freq*sampleRateRec;
+}
+
+void BlendOscillator::setWaveForm1(int newWaveForm1)
+{
+  if( waveTable1 != NULL )
+    waveTable1->setWaveform(newWaveForm1);
+}
+
+void BlendOscillator::setWaveTable1(MipMappedWaveTable* newWaveTable1)
+{
+  waveTable1 = newWaveTable1;
+}
+
+void BlendOscillator::setStartPhase(double StartPhase)
+{
+  if( (StartPhase>=0) && (StartPhase<=360) )
+    startIndex = (StartPhase/360.0)*tableLengthDbl;
+}
+
+//-------------------------------------------------------------------------------------------------
+// event processing:
+
+void BlendOscillator::resetPhase()
+{
+  phaseIndex = startIndex;
+}
+
+void BlendOscillator::setPhase(double PhaseIndex)
+{
+  phaseIndex = startIndex+PhaseIndex;
+}
